+++ src/main/java/org/finalproject/tmeroom/comment/data/entity/Comment.java
@@ -1,43 +1,40 @@
-package org.finalproject.tmeroom.file.data.entity;
+package org.finalproject.tmeroom.comment.data.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.finalproject.tmeroom.common.data.entity.BaseTimeEntity;
-import org.finalproject.tmeroom.lecture.data.entity.Lecture;
-import org.finalproject.tmeroom.member.data.entity.Member;
+import org.finalproject.tmeroom.member.data.Entity.Member;
+import org.finalproject.tmeroom.question.data.Entity.Question;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
 /**
  * 작성자: 김종민
  * 작성일자: 2023-09-11
- * File Entity 작성
+ * Comment Entity 작성
  */
 @Entity
 @Getter
 @RequiredArgsConstructor
-public class File extends BaseTimeEntity {
+public class Comment extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "lecture_id")
-    @NotNull
+    @JoinColumn(name = "commenter_id")
     @OnDelete(action = OnDeleteAction.CASCADE)
-    private Lecture lecture;
+    @NotNull
+    private Member commenter;
 
     @NotNull
-    private String fileName;
+    private String content;
 
+    @ManyToOne
+    @JoinColumn(name = "question_id")
+    @OnDelete(action = OnDeleteAction.CASCADE)
     @NotNull
-    private String fileLink;
-
-    @Column
-    private String uploaderNickname;
-
-    @Column
-    private String fileType;
+    private Question question;
 }
